third_party/blink/web_tests/accessibility/adjacent-continuations-cause-assertion-failure.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>

<div id="container">
  <span><div></div></span><span>x<div>y</div>z</span>
</div>

<p id="description"></p>
<div id="console"></div>

<script>
    description("Make sure that a debug assert is not triggered when constructing the accessibility tree for this page.");

    if (window.accessibilityController) {
        buildAccessibilityTree(accessibilityController.accessibleElementById('container'), 0);
    }

</script>

</body>
</html>
